+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 * Add support for tracing (`micrometer-tracing-bridge-otel`, `opentelemetry-exporter-otlp`)
 * Add Spring Security configuration (`spring-boot-starter-security`) requiring HTTP Basic
   authentication for `/actuator/prometheus`
+* MobileID authentication response validation is optional, disabled by default. New 
+  configuration key `mobileid.validateAuthenticationResponse`
 * Dependency updates
-
 
 ## [0.8.0]